requirements.txt
aiofiles==0.5.0
aiosqlite==0.15.0
alembic==1.4.3
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
cssselect==1.1.0
cssutils==1.0.2
databases==0.3.2
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
emails==0.6
fastapi==0.61.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
orjson==3.4.0
passlib==1.7.2
premailer==3.7.0
promise==2.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
python-dateutil==2.8.1
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
